+++ src/test/kotlin/step3/CarRacingTest.kt
@@ -4,6 +4,8 @@ import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.CsvSource
+import racing.application.CarRacing
+import racing.model.Car
 
 internal class CarRacingTest {
